middle_layer/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
pydantic-settings
slowapi